backend/flask/requirements.txt
attrs==24.2.0
bcrypt==4.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
